+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Remove trailing new line at the end of `chat.md`
 - Fence piped input when there is also a prompt
 - Draw feedback in an input frame box
+- List changed models after a startup refresh
 
 ## [0.6.0] - 2026-09-18